requirements.txt
appnope==0.1.0
backcall==0.1.0
confusable-homoglyphs==3.0.0
coverage==4.5.1
decorator==4.3.0
Django==2.0.4
django-multiselectfield==0.1.8
django-registration==2.4.1
factory-boy==2.10.0
Faker==0.8.13
flake8==3.5.0
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
mccabe==0.6.1
model-mommy==1.5.1
parso==0.2.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.7.2
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
